Separate config parser from Pipeline class
