+++ skypy/pipeline/__init__.py
@@ -3,4 +3,5 @@ This module provides methods to pipeline together multiple models with
 dependencies and handle their outputs.
 """
 
+from ._config import *  # noqa
 from ._pipeline import *  # noqa
